Renamed imperial.js -> ImperialGameCoordinator
